+++ test/test_tae/test_exec_tae_run.py
@@ -84,7 +84,7 @@ class TaeTest(unittest.TestCase):
                                   'output_dir': ''}, cmd_args=None)
         stats = Stats(scen)
         stats.start_timing()
-        eta = ExecuteTARun(ta=lambda *args: None, stats=stats)
+        eta = ExecuteTARun(ta=lambda *args: None, stats=stats, run_obj="quality")
 
         self.assertRaises(
             FirstRunCrashedException, eta.start, config={}, instance=1)